kernel/mqueue: Match types of variable between structures
- Type of variables about message length, 'size_t'
   . maxmsgsize in mqueue_inode_s
   . msglen in mqueue_msg_s
   . msglen parameters of MQ functions.

- Type of variables about message count, 'uint16_t'
   . maxmsgs, nmsgs in mqueue_inode_s
   . mq_maxmsg, mq_msgsize in mq_attr.

+++ os/include/mqueue.h
@@ -85,10 +85,10 @@
 
 /** @brief sturcutre of mqueue attritube */
 struct mq_attr {
-	size_t mq_maxmsg;			/* Max number of messages in queue */
+	uint16_t mq_maxmsg;			/* Max number of messages in queue */
 	size_t mq_msgsize;			/* Max message size */
 	unsigned mq_flags;			/* Queue flags */
-	size_t mq_curmsgs;			/* Number of messages currently in queue */
+	uint16_t mq_curmsgs;			/* Number of messages currently in queue */
 };
 
 /* Message queue descriptor */
